chore: resolve comments from the child PR
