+++ packages/relay/src/lib/types/index.ts
@@ -13,3 +13,4 @@ export * from './spendingPlanConfig';
 export * from './registry';
 export * from './debug';
 export * from './rateLimiter';
+export * from './sdkClient';
